submissions/team-members/david-boules/requirements.txt
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
split-folders>=0.5.1
tensorflow>=2.12.0
Pillow>=9.5.0
jupyter>=1.0.0
ipython>=8.0.0
ipykernel>=6.0.0
scipy>=1.10.0
scikit-learn>=1.0.2
seaborn>=0.12.0
